updatign references
